+++ spec2vec/__init__.py
@@ -1,7 +1,6 @@
 from .__version__ import __version__
 from .Document import Document
 from .Spec2Vec import Spec2Vec
-from .Spec2VecParallel import Spec2VecParallel
 from .SpectrumDocument import SpectrumDocument
 from .vector_operations import calc_vector
 
@@ -12,5 +11,4 @@ __all__ = [
     "Document",
     "SpectrumDocument",
     "Spec2Vec",
-    "Spec2VecParallel"
 ]