+++ CHANGELOG.md
@@ -11,6 +11,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Implement faster, numba-based cosine similarity function [#29](https://github.com/iomega/spec2vec/pull/29)
 
+## [0.3.0] - 2020-09-16
+
+### Added
+
+- Support for Python 3.8 [#35](https://github.com/iomega/spec2vec/pull/35)
+
+### Changed
+
+- Refactored Spec2Vec class to provide .pair() and .matrix() methods [#35](https://github.com/iomega/spec2vec/pull/35)
+
+### Removed
+
+- Spec2VecParallel (is now included as Spec2Vec.matrix()) [#35](https://github.com/iomega/spec2vec/pull/35)
+
 ## [0.2.0] - 2020-06-18
 
 ### Added
@@ -48,6 +62,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fossa configuration
 - Flowchart
 
-[Unreleased]: https://github.com/iomega/spec2vec/compare/0.2.0...HEAD
+[Unreleased]: https://github.com/iomega/spec2vec/compare/0.3.0...HEAD
+[0.3.0]: https://github.com/iomega/spec2vec/releases/tag/0.3.0
 [0.2.0]: https://github.com/iomega/spec2vec/releases/tag/0.2.0
 [0.1.0]: https://github.com/iomega/spec2vec/releases/tag/0.1.0